move useTable hooks, remove default sort
